add sweatalert2

+++ src/functions/fetchData.js
@@ -64,17 +64,26 @@ export const getTodoData = async (userId) => {
 
 export function deleteTodo(){
     const todoId = this.id.split('-')[1];
-    fetch(`/api/todos/${todoId}`, {
-        method: 'DELETE'
-    })
-    .then((res) => {
-        if(res.status === 200){
-            alert('Delete todo successfully');
-            window.location.reload();
-        } else {
-            throw new Error('Failed to delete todo');
+    swal({
+        title: "Do you want to delete the task?",
+        showCancelButton: true,
+        confirmButtonText: `Delete`,
+    }).then((result) => {
+        if (result.value) {
+            return fetch(`/api/todos/${todoId}`, {
+                method: 'DELETE'
+            })
+        } 
+    }).then((res) => {
+        if (res) {
+            if(res.status === 200){
+                return swal('<h3 class="font-bold text-xl">Task deleted!</h3>', '', 'success')
+            } else {
+                throw new Error('Failed to delete todo');
+            }
         }
     })
+    .then(() => window.location.reload())
     .catch(err => console.error(err));
 }
 
@@ -85,7 +94,7 @@ export function updateTodoStatus(){
     })
     .then((res) => {
         if(res.status === 200){
-            alert('Update todo status successfully');
+            swal('<h3 class="font-bold text-xl">Update todo status successfully</h3>','','success');
         } else {
             throw new Error('Failed to update todo status');
         }
